Move SlackManager to slack.py,move token to config.ini, instance rq_decode with params
