+++ development.ini
@@ -14,6 +14,12 @@ pyramid.default_locale_name = en
 pyramid.includes =
     pyramid_debugtoolbar
 
+device_checkout_token = WyQMp2xxs4hXTpEXM43SOLzM
+device_list_token = Bnh4eVVru7Xmf0WUI0a8APqs
+device_reg_token = cJlNFZt2cRgehBPiQq0vD5g9
+device_dereg_token = jB6UIvVFgACe18YPHltpqqPj
+device_audit_token = v6X6evAJ4pDcSxc0feKLxQ6D
+
 # By default, the toolbar only appears for clients from IP addresses
 # '127.0.0.1' and '::1'.
 # debugtoolbar.hosts = 127.0.0.1 ::1
@@ -25,7 +31,7 @@ pyramid.includes =
 [server:main]
 use = egg:waitress#main
 host = 0.0.0.0
-port = %(http_port)s
+port = 6543
 
 [alembic]
 script_location = genius_bar/alembic
@@ -37,7 +43,7 @@ sqlalchemy.url = mysql+pymysql://root@localhost:3306/genius_bar_test?charset=utf
 ###
 
 [loggers]
-keys = root, tweetmon, sqlalchemy, alembic
+keys = root, sqlalchemy, alembic
 
 [handlers]
 keys = console
@@ -48,11 +54,6 @@ keys = generic
 [logger_root]
 level = INFO
 handlers = console
-
-[logger_tweetmon]
-level = DEBUG
-handlers =
-qualname = tweetmon
 
 [logger_sqlalchemy]
 level = WARN
